Modify ParseSet() to take a filename argument
A case exists where a developer may want to parse a set of flags from a
configuration file that is not specified via a flag. Currently,
flagconfig forces the user to use the '-c' flag. This commit modifies
ParseSet() to take a filename argument so that an application can
specify a specific file path programmatically.